+++ src/plonemeeting/portal/core/browser/tests/test_institution.py
@@ -1,0 +1,34 @@
+# -*- coding: utf-8 -*-
+
+from plonemeeting.portal.core.testing import PLONEMEETING_PORTAL_DEMO_FUNCTIONAL_TESTING
+from plone.app.testing import login
+from plone.app.testing import logout
+from plone.app.testing import setRoles
+from plone.app.testing import TEST_USER_ID
+from plone.app.testing import TEST_USER_NAME
+
+import unittest
+
+
+class TestInstitutionView(unittest.TestCase):
+    layer = PLONEMEETING_PORTAL_DEMO_FUNCTIONAL_TESTING
+
+    def setUp(self):
+        self.portal = self.layer["portal"]
+
+    def tearDown(self):
+        logout()
+
+    def test_call_manager(self):
+        institution = self.portal["belleville"]
+        setRoles(self.portal, TEST_USER_ID, ["Manager"])
+        login(self.portal, TEST_USER_NAME)
+        view_content = institution.restrictedTraverse("@@view")()
+        self.assertTrue("Meeting config ID" in view_content)
+
+    def test_call_anonymous(self):
+        institution = self.portal["belleville"]
+        setRoles(self.portal, TEST_USER_ID, ["Reader"])
+        login(self.portal, TEST_USER_NAME)
+        view_content = institution.restrictedTraverse("@@view")()
+        self.assertTrue("Meeting config ID" not in view_content)